feature/user-interfaces (#24)
1. Не исправлены ошибки линтера
2. Не исправлены интеграционные тесты

* feature/user-interfaces: base react app with auth and license creation

* feature/user-interfaces: upd mocks

* feature/user-interfaces: controllers

* feature/user-interfaces: MONSTER COMMIT. IDGAF. no time

* feature/user-interfaces: updated some shit

* feature/user-interfaces(web): updated admin panel

* feature/user-interfaces(cli): done

* feature/user-interfaces(web): updated logo

* feature/user-interfaces: fixed search and some shi

* feature/user-interfaces(cli): deleted backend readme

* feature/user-interfaces(cli): added track stats in cli (without stat updating)

* feature/user-interfaces(cli): cmdrouter moved into pkg

* feature/user-interfaces(cli): updated cli, without track seek and stats

* feature/user-interfaces(cli): updated domain with tests

* feature/user-interfaces(cli): updated repositories, added change_track_position SP

* feature/user-interfaces(cli): fixed deletion track from playlist

* feature/user-interfaces(cli): fixed integration test

* feature/user-interfaces(domain): added claims nil checking ^_^

* feature/user-interfaces(http): http controllers layer is done. ['_']

* feature/user-interfaces: updated env example

* feature/user-interfaces: added json tags

* feature/user-interfaces: upd dependencies

* feature/user-interfaces(web): tempo

* feature/user-interfaces(web): tempo with fworking home page and album page

* feature/user-interfaces(http ctrl): some fixes

* feature/user-interfaces(db): fixes in db scheme

* feature/user-interfaces(tests): fixed integration tests

* feature/user-interfaces: upd dependencies

* feature/user-interfaces(web): added playlist page

* feature/user-interfaces: deleted coverage report

* feature/user-interfaces: fixed mocks

* feature/user-interfaces: fixed tests

* feature/user-interfaces: fixed genre meta tests

* feature/user-interfaces: test music

* feature/user-interfaces: upd dependencies

* feature/user-interfaces: fixed tracks addition

* feature/user-interfaces: upd db scheme

* feature/user-interfaces(web): upd playlists presentation

* feature/user-interfaces(backend): added aggregators

* feature/user-interfaces(backend): upd db scheme

* feature/user-interfaces(web): upd playlists and albums

* feature/user-interfaces(web): upd front with artist profil

* feature/user-interfaces(web): upd search

* feature/user-interfaces(web): upd profile

* feature/user-interfaces(backend): updated policy check

* feature/user-interfaces(backend): updated db

* feature/user-interfaces(backend): fixed updated_at

* feature/user-interfaces(backend): fixed some shit

* feature/user-interfaces(backend): updated frontend. tried solid

* feature/user-interfaces(uml): upd

* feature/user-interfaces: updated mocks

* feature/user-interfaces: updated tests

* feature/user-interfaces: upd github workflow

* feature/user-interfaces: deleted flag for golangcli-lint

* feature/user-interfaces: try to fix linter

* feature/user-interfaces: try to fix linter 2

* feature/user-interfaces: try to fix linter 3

* feature/user-interfaces: try to fix linter 4

+++ backend/internal/domain/services/content/album/cover/cover.go
@@ -2,16 +2,12 @@ package cover
 
 import (
 	"context"
-	"errors"
 
 	"github.com/google/uuid"
 	"github.com/hahaclassic/orpheon/backend/internal/domain/entity"
 	usecase "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/content/album"
+	commonerr "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/errors"
 	"github.com/hahaclassic/orpheon/backend/pkg/errwrap"
-)
-
-var (
-	ErrForbidden = errors.New("permission denied")
 )
 
 type AlbumCoverRepository interface {
@@ -43,8 +39,8 @@ func (c *AlbumCoverService) UploadCover(ctx context.Context, claims *entity.Clai
 		err = errwrap.WrapIfErr(usecase.ErrUploadCover, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return c.repo.SaveCover(ctx, cover)
@@ -55,8 +51,8 @@ func (c *AlbumCoverService) DeleteCover(ctx context.Context, claims *entity.Clai
 		err = errwrap.WrapIfErr(usecase.ErrDeleteCover, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return c.repo.DeleteCover(ctx, albumID)
