+++ backend/internal/domain/services/content/album/meta/meta.go
@@ -7,12 +7,12 @@ import (
 	"github.com/google/uuid"
 	"github.com/hahaclassic/orpheon/backend/internal/domain/entity"
 	usecase "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/content/album"
+	commonerr "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/errors"
 	"github.com/hahaclassic/orpheon/backend/pkg/errwrap"
 )
 
 var (
 	ErrGenerateID = errors.New("id generation error")
-	ErrForbidden  = errors.New("permission denied error")
 )
 
 type AlbumRepository interface {
@@ -20,6 +20,9 @@ type AlbumRepository interface {
 	GetAlbum(ctx context.Context, id uuid.UUID) (*entity.AlbumMeta, error)
 	UpdateAlbum(ctx context.Context, album *entity.AlbumMeta) error
 	DeleteAlbum(ctx context.Context, id uuid.UUID) error
+	GetAllAlbums(ctx context.Context) ([]*entity.AlbumMeta, error)
+	GetAlbumArtists(ctx context.Context, albumID uuid.UUID) ([]*entity.ArtistMeta, error)
+	GetAlbumGenres(ctx context.Context, albumID uuid.UUID) ([]*entity.Genre, error)
 }
 
 type AlbumService struct {
@@ -32,20 +35,25 @@ func New(repo AlbumRepository) *AlbumService {
 	}
 }
 
-func (a *AlbumService) CreateAlbum(ctx context.Context, claims *entity.Claims, album *entity.AlbumMeta) (err error) {
+func (a *AlbumService) CreateAlbum(ctx context.Context, claims *entity.Claims, album *entity.AlbumMeta) (id uuid.UUID, err error) {
 	defer func() {
 		err = errwrap.WrapIfErr(usecase.ErrCreateAlbum, err)
 	}()
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return uuid.Nil, commonerr.ErrForbidden
 	}
 
 	album.ID, err = uuid.NewRandom()
 	if err != nil {
-		return ErrGenerateID
+		return uuid.Nil, ErrGenerateID
 	}
 
-	return a.repo.CreateAlbum(ctx, album)
+	err = a.repo.CreateAlbum(ctx, album)
+	if err != nil {
+		return uuid.Nil, err
+	}
+
+	return album.ID, nil
 }
 
 func (a *AlbumService) GetAlbum(ctx context.Context, albumID uuid.UUID) (_ *entity.AlbumMeta, err error) {
@@ -56,13 +64,21 @@ func (a *AlbumService) GetAlbum(ctx context.Context, albumID uuid.UUID) (_ *enti
 	return a.repo.GetAlbum(ctx, albumID)
 }
 
+func (a *AlbumService) GetAllAlbums(ctx context.Context) (_ []*entity.AlbumMeta, err error) {
+	defer func() {
+		err = errwrap.WrapIfErr(usecase.ErrGetAllAlbums, err)
+	}()
+
+	return a.repo.GetAllAlbums(ctx)
+}
+
 func (a *AlbumService) UpdateAlbum(ctx context.Context, claims *entity.Claims, album *entity.AlbumMeta) (err error) {
 	defer func() {
 		err = errwrap.WrapIfErr(usecase.ErrUpdateAlbum, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return a.repo.UpdateAlbum(ctx, album)
@@ -73,8 +89,8 @@ func (a *AlbumService) DeleteAlbum(ctx context.Context, claims *entity.Claims, a
 		err = errwrap.WrapIfErr(usecase.ErrDeleteAlbum, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return a.repo.DeleteAlbum(ctx, albumID)
