+++ backend/internal/domain/services/content/artist/avatar/avatar.go
@@ -2,16 +2,12 @@ package avatar
 
 import (
 	"context"
-	"errors"
 
 	"github.com/google/uuid"
 	"github.com/hahaclassic/orpheon/backend/internal/domain/entity"
 	usecase "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/content/artist"
+	commonerr "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/errors"
 	"github.com/hahaclassic/orpheon/backend/pkg/errwrap"
-)
-
-var (
-	ErrForbidden = errors.New("permission denied")
 )
 
 type ArtistAvatarRepository interface {
@@ -43,8 +39,8 @@ func (s *ArtistCoverService) UploadCover(ctx context.Context, claims *entity.Cla
 		err = errwrap.WrapIfErr(usecase.ErrUploadAvatar, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return s.repo.SaveCover(ctx, cover)
@@ -55,8 +51,8 @@ func (s *ArtistCoverService) DeleteCover(ctx context.Context, claims *entity.Cla
 		err = errwrap.WrapIfErr(usecase.ErrDeleteAvatar, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return s.repo.DeleteCover(ctx, artistID)
