+++ backend/internal/domain/services/content/track/meta/meta.go
@@ -6,13 +6,14 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/hahaclassic/orpheon/backend/internal/domain/entity"
+	"github.com/hahaclassic/orpheon/backend/internal/domain/usecases/content/track"
 	usecase "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/content/track"
+	commonerr "github.com/hahaclassic/orpheon/backend/internal/domain/usecases/errors"
 	"github.com/hahaclassic/orpheon/backend/pkg/errwrap"
 )
 
 var (
 	ErrGenerateTrackID = errors.New("generate track id error")
-	ErrForbidden       = errors.New("permission denied")
 )
 
 type TrackMetaRepository interface {
@@ -23,11 +24,12 @@ type TrackMetaRepository interface {
 }
 
 type TrackMetaService struct {
-	repo TrackMetaRepository
+	repo           TrackMetaRepository
+	segmentService track.TrackSegmentService
 }
 
-func NewTrackMetaService(repo TrackMetaRepository) *TrackMetaService {
-	return &TrackMetaService{repo: repo}
+func NewTrackMetaService(repo TrackMetaRepository, segmentService track.TrackSegmentService) *TrackMetaService {
+	return &TrackMetaService{repo: repo, segmentService: segmentService}
 }
 
 func (s *TrackMetaService) GetTrackMeta(ctx context.Context, trackID uuid.UUID) (_ *entity.TrackMeta, err error) {
@@ -35,16 +37,21 @@ func (s *TrackMetaService) GetTrackMeta(ctx context.Context, trackID uuid.UUID) 
 		err = errwrap.WrapIfErr(usecase.ErrGetTrackMeta, err)
 	}()
 
-	return s.repo.GetByID(ctx, trackID)
+	track, err := s.repo.GetByID(ctx, trackID)
+	if err != nil {
+		return nil, err
+	}
+
+	return track, nil
 }
 
-func (s *TrackMetaService) CreateTrackMeta(ctx context.Context, claims *entity.Claims, track *entity.TrackMeta) (id uuid.UUID, err error) {
+func (s *TrackMetaService) CreateTrackMeta(ctx context.Context, claims *entity.Claims, track *entity.TrackMeta) (_ uuid.UUID, err error) {
 	defer func() {
 		err = errwrap.WrapIfErr(usecase.ErrCreateTrackMeta, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return uuid.Nil, ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return uuid.Nil, commonerr.ErrForbidden
 	}
 
 	track.ID, err = uuid.NewRandom()
@@ -56,7 +63,11 @@ func (s *TrackMetaService) CreateTrackMeta(ctx context.Context, claims *entity.C
 		return uuid.Nil, err
 	}
 
-	return id, nil
+	if err = s.segmentService.CreateSegments(ctx, track.ID, track.Duration); err != nil {
+		return uuid.Nil, err
+	}
+
+	return track.ID, nil
 }
 
 func (s *TrackMetaService) UpdateTrackMeta(ctx context.Context, claims *entity.Claims, track *entity.TrackMeta) (err error) {
@@ -64,8 +75,8 @@ func (s *TrackMetaService) UpdateTrackMeta(ctx context.Context, claims *entity.C
 		err = errwrap.WrapIfErr(usecase.ErrUpdateTrackMeta, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
 	}
 
 	return s.repo.Update(ctx, track)
@@ -76,8 +87,12 @@ func (s *TrackMetaService) DeleteTrackMeta(ctx context.Context, claims *entity.C
 		err = errwrap.WrapIfErr(usecase.ErrDeleteTrackMeta, err)
 	}()
 
-	if claims.AccessLvl != entity.Admin {
-		return ErrForbidden
+	if claims == nil || claims.AccessLvl != entity.Admin {
+		return commonerr.ErrForbidden
+	}
+
+	if err = s.segmentService.DeleteSegments(ctx, trackID); err != nil {
+		return err
 	}
 
 	return s.repo.Delete(ctx, trackID)
